Fixes error in restoring jump routes
The method that reconfigured the jump routes after reading them
was attempting to make a function call off a pointer that didn't
actually point anywhere but was temporarily storing the ID value
of the system it should point to.  The call was fixed but due
to size differences required adding the -fpermissive flag
to the compile command.