Fix drawing of background in main frame of strategic game
This fixes the problem of not being able to paint a background in
the main frame by adding a panel that is drawn in.  Only fixes the
main strategic game, not the tactical game board.

Note: at this point, clicks in the new panel are not being registered
and acted upon.